Create subclass for transformer state and shared functions.
