+++ lib/Compressor.js
@@ -5,13 +5,14 @@ import * as cborg from 'cborg';
 import {CborldError} from './CborldError.js';
 import {CborldEncoder} from './codecs/CborldEncoder.js';
 import {ContextEncoder} from './codecs/ContextEncoder.js';
+import {Transformer} from './Transformer.js';
 import {UriEncoder} from './codecs/UriEncoder.js';
 // FIXME: make extensible
 import {VocabTermEncoder} from './codecs/VocabTermEncoder.js';
 import {XsdDateEncoder} from './codecs/XsdDateEncoder.js';
 import {XsdDateTimeEncoder} from './codecs/XsdDateTimeEncoder.js';
 import {inspect} from './util.js';
-import {getActiveContext} from './context2.js';
+import {KEYWORDS} from './keywords.js';
 
 // FIXME: move elsewhere
 const TYPE_ENCODERS = new Map([
@@ -24,41 +25,6 @@ const TYPE_ENCODERS = new Map([
 
 const CONTEXT_TERM_ID = 0;
 
-// FIXME: move elsewhere
-const KEYWORDS = new Map([
-  // ordered is important, do not change
-  // FIXME: can we alphabetize these? do they all fit into whatever varint max
-  // size thing is applied?
-  ['@context', 0],
-  ['@type', 2],
-  ['@id', 4],
-  ['@value', 6],
-  ['@base', 8],
-  ['@language', 10],
-  ['@direction', 12],
-  ['@graph', 14],
-  ['@included', 16],
-  ['@json', 18],
-  ['@list', 20],
-  ['@set', 22],
-  ['@index', 24],
-  ['@nest', 26],
-  ['@reverse', 28],
-  // FIXME: remove these? these only appear in frames and contexts
-  ['@container', 30],
-  ['@default', 32],
-  ['@embed', 34],
-  ['@explicit', 36],
-  ['@none', 38],
-  ['@omitDefault', 40],
-  ['@prefix', 42],
-  ['@preserve', 44],
-  ['@protected', 46],
-  ['@requireAll', 48],
-  ['@version', 50],
-  ['@vocab', 52]
-]);
-
 const COMPRESSED_PREFIX = new Uint8Array([0xd9, 0x05, 0x01]);
 
 // override cborg object encoder to use cborld encoders
@@ -70,7 +36,7 @@ const typeEncoders = {
   }
 };
 
-export class Compressor {
+export class Compressor extends Transformer {
   /**
    * Creates a new Compressor for generating compressed CBOR-LD from a
    * JSON-LD document. The created instance may only be used on a single
@@ -84,9 +50,7 @@ export class Compressor {
    *   (0x7FFF)).
    */
   constructor({appContextMap, documentLoader} = {}) {
-    // FIXME: rename `appContextMap`
-    this.appContextMap = appContextMap;
-    this.documentLoader = documentLoader;
+    super({appContextMap, documentLoader});
   }
 
   /**
@@ -152,8 +116,7 @@ export class Compressor {
     const {termToId, appContextMap} = this;
 
     // get active context for this object
-    const activeCtx = await getActiveContext(
-      {obj, contextStack, transformer: this});
+    const activeCtx = await this._getActiveContext({obj, contextStack});
     const {childContextStack} = activeCtx;
 
     // if `@context` is present in the object, encode it first
@@ -193,7 +156,7 @@ export class Compressor {
       let values = obj[key];
       const isArray = Array.isArray(values);
 
-      const termId = _getIdForTerm({term: key, termToId, plural: isArray});
+      const termId = this._getIdForTerm({term: key, plural: isArray});
       console.log('termId', termId);
 
       // handle `@id`, which must have a single value
@@ -235,7 +198,7 @@ export class Compressor {
       const propertyScopedContext = scopedContextMap.get(key);
       if(propertyScopedContext) {
         // TODO: support `@propagate: false` on property-scoped contexts
-        newActiveCtx = await getActiveContext({
+        newActiveCtx = await this._getActiveContext({
           obj: {'@context': propertyScopedContext},
           contextStack: childContextStack, scoped: true
         });
@@ -249,7 +212,7 @@ export class Compressor {
         }
 
         // try to get a type-based encoder for the value
-        const termType = _getTermType(
+        const termType = this._getTermType(
           {activeCtx: newActiveCtx || activeCtx, def});
         console.log('termType', termType);
         const EncoderClass = TYPE_ENCODERS.get(termType);
@@ -301,44 +264,6 @@ export class Compressor {
     // revert active context for this object
     activeCtx.revert();
   }
-}
-
-function _getTermType({activeCtx, def}) {
-  const {'@type': type} = def;
-  if(!type) {
-    // no term type
-    return;
-  }
-
-  // check for potential CURIE value
-  const [prefix, ...suffix] = type.split(':');
-  const prefixDef = activeCtx.termMap.get(prefix);
-  if(prefixDef === undefined) {
-    // no CURIE
-    return type;
-  }
-
-  // handle CURIE
-  if(typeof prefixDef === 'string') {
-    return prefixDef + suffix.join(':');
-  }
-
-  // prefix definition must be an object
-  if(!(typeof prefixDef === 'object' &&
-    typeof prefixDef['@id'] === 'string')) {
-    // FIXME: use CborldError
-    throw new Error('Invalid term definition');
-  }
-  return prefixDef['@id'] + suffix.join(':');
-}
-
-function _getIdForTerm({term, termToId, plural}) {
-  const id = termToId.get(term);
-  if(id === undefined) {
-    // FIXME: use CBOR-LD error
-    throw new Error('term not defined');
-  }
-  return plural ? id + 1 : id;
 }
 
 /**
