+++ lib/Decompressor.js
@@ -3,13 +3,14 @@
  */
 import * as cborg from 'cborg';
 import {CborldError} from './CborldError.js';
-import {getActiveContext} from './context2.js';
 import {ContextDecoder} from './codecs/ContextDecoder.js';
+import {Transformer} from './Transformer.js';
 // FIXME: make extensible
 import {UriDecoder} from './codecs/UriDecoder.js';
 import {VocabTermDecoder} from './codecs/VocabTermDecoder.js';
 import {XsdDateDecoder} from './codecs/XsdDateDecoder.js';
 import {XsdDateTimeDecoder} from './codecs/XsdDateTimeDecoder.js';
+import {KEYWORDS} from './keywords.js';
 
 // FIXME: move elsewhere
 // const COMPRESSED_PREFIX = new Uint8Array([0xd9, 0x05, 0x01]);
@@ -27,64 +28,29 @@ const TYPE_DECODERS = new Map([
 // FIXME: move elsewhere
 const CONTEXT_TERM_ID = 0;
 
-// FIXME: move elsewhere
-const KEYWORDS = new Map([
-  // ordered is important, do not change
-  // FIXME: can we alphabetize these? do they all fit into whatever varint max
-  // size thing is applied?
-  ['@context', 0],
-  ['@type', 2],
-  ['@id', 4],
-  ['@value', 6],
-  ['@base', 8],
-  ['@language', 10],
-  ['@direction', 12],
-  ['@graph', 14],
-  ['@included', 16],
-  ['@json', 18],
-  ['@list', 20],
-  ['@set', 22],
-  ['@index', 24],
-  ['@nest', 26],
-  ['@reverse', 28],
-  // FIXME: remove these? these only appear in frames and contexts
-  ['@container', 30],
-  ['@default', 32],
-  ['@embed', 34],
-  ['@explicit', 36],
-  ['@none', 38],
-  ['@omitDefault', 40],
-  ['@prefix', 42],
-  ['@preserve', 44],
-  ['@protected', 46],
-  ['@requireAll', 48],
-  ['@version', 50],
-  ['@vocab', 52]
-]);
-
-export class Decompressor {
+export class Decompressor extends Transformer {
   /**
    * Creates a new Decompressor for generating a JSON-LD document from
    * compressed CBOR-LD. The created instance may only be used on a single
    * CBOR-LD input at a time.
    *
    * @param {object} options - The options to use when encoding to CBOR-LD.
-   * @param {documentLoaderFunction} options.documentLoader -The document
+   * @param {documentLoaderFunction} options.documentLoader - The document
    *   loader to use when resolving JSON-LD Context URLs.
-   * // FIXME: this needs to be a reverse map... or we need to build it
    * @param {Map} [options.appContextMap] - A map of JSON-LD Context URLs and
    *   their encoded CBOR-LD values (must be values greater than 32767
    *   (0x7FFF)).
    */
   constructor({appContextMap, documentLoader} = {}) {
+    super({appContextMap, documentLoader});
     // FIXME: rename `appContextMap`
     this.appContextMap = new Map();
+    // FIXME: we need to pass a reverse map... or we need to build it
     if(appContextMap) {
       for(const [k, v] of appContextMap) {
         this.appContextMap.set(v, k);
       }
     }
-    this.documentLoader = documentLoader;
   }
 
   /**
@@ -174,8 +140,8 @@ export class Decompressor {
     }
 
     // get active context for this object
-    const activeCtx = await getActiveContext(
-      {obj, encodedMap, contextStack, transformer: this});
+    const activeCtx = await this._getActiveContext(
+      {obj, encodedMap, contextStack});
     const {childContextStack} = activeCtx;
 
     // walk the encoded map
@@ -187,7 +153,7 @@ export class Decompressor {
       }
 
       // check for undefined term IDs
-      const {term, plural} = _getTermForId({id: key, idToTerm});
+      const {term, plural} = this._getTermForId({id: key});
       if(term === undefined) {
         throw new CborldError(
           'ERR_UNKNOWN_CBORLD_TERM_ID',
@@ -234,9 +200,9 @@ export class Decompressor {
       const propertyScopedContext = scopedContextMap.get(key);
       if(propertyScopedContext) {
         // TODO: support `@propagate: false` on property-scoped contexts
-        newActiveCtx = await getActiveContext({
+        newActiveCtx = await this._getActiveContext({
           obj: {'@context': propertyScopedContext},
-          contextStack: childContextStack, transformer: this, scoped: true
+          contextStack: childContextStack, scoped: true
         });
       }
 
@@ -248,7 +214,7 @@ export class Decompressor {
         }
 
         // try to get a type-based decoder for the value
-        const termType = _getTermType(
+        const termType = this._getTermType(
           {activeCtx: newActiveCtx || activeCtx, def});
         console.log('termType', termType);
         const DecoderClass = TYPE_DECODERS.get(termType);
@@ -300,41 +266,6 @@ export class Decompressor {
     // revert active context for this object
     activeCtx.revert();
   }
-}
-
-function _getTermType({activeCtx, def}) {
-  const {'@type': type} = def;
-  if(!type) {
-    // no term type
-    return;
-  }
-
-  // check for potential CURIE value
-  const [prefix, ...suffix] = type.split(':');
-  const prefixDef = activeCtx.termMap.get(prefix);
-  if(prefixDef === undefined) {
-    // no CURIE
-    return type;
-  }
-
-  // handle CURIE
-  if(typeof prefixDef === 'string') {
-    return prefixDef + suffix.join(':');
-  }
-
-  // prefix definition must be an object
-  if(!(typeof prefixDef === 'object' &&
-    typeof prefixDef['@id'] === 'string')) {
-    // FIXME: use CborldError
-    throw new Error('Invalid term definition');
-  }
-  return prefixDef['@id'] + suffix.join(':');
-}
-
-function _getTermForId({id, idToTerm}) {
-  const plural = (id & 1) === 1;
-  const term = idToTerm.get(plural ? id - 1 : id);
-  return {term, plural};
 }
 
 /**
